Codecov test.

+++ music/convert.sh
@@ -9,6 +9,7 @@ if [[ $*\  == *--help\ * || $*\  == *-h\ * ]]; then
     echo -e "\nFlaggor:"
     echo -e "\t-c, --compress\tKomprimera svg-filerna"
     echo -e "\t-f, --force\tTvinga omgenerering av existerande filer"
+    echo -e "\t--no-generate\tGenerara inga nya noter"
     echo -e "\t--no-hash\tInkludera inte en checksum i filnamnet"
     echo -e "\t--no-json\tStrunta i att generera svgs.json"
     echo -e "\t--no-move\tFlytta inte filerna till deras målmappar (för användning med Vue)"
@@ -30,49 +31,52 @@ fi
 export QT_QPA_PLATFORM=offscreen
 SCALE_FACTORS=(1 1.25 1.5 1.75 2 3 4)
 
-echo -e "\e[1;32mGenererar SVG-filer:\e[0m"
-for path in mscz/*.mscz
-do
-    file=$(basename "$path")
-    echo -e " - \e[34m$file\e[0m"
-
-    # Clear old files in tmp dir.
-    rm tmp/*.mscx
-
-    # Extract the source file, so that we can modify the scale factor before rendering.
-    # Note that the internal .mscx file does not neccessarily correspond to the .mscz file name if the .mscz has been renamed after saving.
-    unzip -o "mscz/$file" *.mscx -d "tmp" || exit 3;
-
-    for sf in "${SCALE_FACTORS[@]}"
+if [[ ! $*\  == *--no-generate\ * ]]; then
+    echo -e "\e[1;32mGenererar SVG-filer:\e[0m"
+    for path in mscz/*.mscz
     do
-        if [[ $*\  == *--no-hash\ * ]]; then
-            # Set hash to empty.
-            csum=""
-        else
-            # Check if another file with the same checksum exists.
-            # If it's found, use the old file.
-            csum="."$(sha1sum "mscz/$file")
-            if ls svg/*${csum:0:8}-sf$sf-*.svg > /dev/null 2>&1
-            then
-                continue
-            fi
-        fi
-        
-        for tmpfile in tmp/*.mscx # Find whatever the .mscx file is called (there should only be one).
+        file=$(basename "$path")
+        echo -e " - \e[34m$file\e[0m"
+
+        # Clear old files in tmp dir.
+        rm tmp/*.mscx
+
+        # Extract the source file, so that we can modify the scale factor before rendering.
+        # Note that the internal .mscx file does not neccessarily correspond to the .mscz file name if the .mscz has been renamed after saving.
+        unzip -o "mscz/$file" *.mscx -d "tmp" || exit 3;
+
+        for sf in "${SCALE_FACTORS[@]}"
         do
-            # Modify the scaling factor and the page height in order to get the desired zoom level in a single svg file.
-            sed -e "s/<Spatium>[0-9]\.[0-9]*<\/Spatium>/<Spatium>$sf<\/Spatium>/g" "$tmpfile" \
-                | sed -e "s/<pageHeight>[0-9]*\.[0-9]*<\/pageHeight>/<pageHeight>1000<\/pageHeight>/g" \
-                > "tmp/${file//mscz/tmp.mscx}" \
-                || exit 3
-            # Generate svg
-            mscore3 --export-to "svg/${file//.mscz/}${csum:0:8}-sf$sf.svg" "tmp/${file//mscz/tmp.mscx}" --force --trim-image 140 || exit 1;
+            if [[ $*\  == *--no-hash\ * ]]; then
+                # Set hash to empty.
+                csum=""
+            else
+                # Check if another file with the same checksum exists.
+                # If it's found, use the old file.
+                csum="."$(sha1sum "mscz/$file")
+                if ls svg/*${csum:0:8}-sf$sf-*.svg > /dev/null 2>&1
+                then
+                    continue
+                fi
+            fi
+            
+            for tmpfile in tmp/*.mscx # Find whatever the .mscx file is called (there should only be one).
+            do
+                # Modify the scaling factor and the page height in order to get the desired zoom level in a single svg file.
+                sed -e "s/<Spatium>[0-9]\.[0-9]*<\/Spatium>/<Spatium>$sf<\/Spatium>/g" "$tmpfile" \
+                    | sed -e "s/<pageHeight>[0-9]*\.[0-9]*<\/pageHeight>/<pageHeight>1000<\/pageHeight>/g" \
+                    > "tmp/${file//mscz/tmp.mscx}" \
+                    || exit 3
+                # Generate svg
+                mscore3 --export-to "svg/${file//.mscz/}${csum:0:8}-sf$sf.svg" "tmp/${file//mscz/tmp.mscx}" --force --trim-image 140 || exit 1;
+            done
         done
     done
-done
+fi
 
 # Cleanup
 rm -rf tmp
+
 
 ##
 ## Compress svg:s
